[template] Contract consolidate txdata error handling
